requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
dj-rest-auth==7.0.1
Django==5.2.1
django-allauth==65.8.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
mysqlclient==2.2.7
packaging==25.0
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
